profile model when creating a model from code
