+++ packages/client/hmi-client/src/workflow/ops/model-from-code/model-from-code-operation.ts
@@ -14,7 +14,10 @@ export const ModelFromCodeOperation: Operation = {
 	description: 'Create model',
 	displayName: 'Create model from code',
 	isRunnable: true,
-	inputs: [{ type: 'codeAssetId', label: 'Code' }],
+	inputs: [
+		{ type: 'codeAssetId', label: 'Code' },
+		{ type: 'documentId', label: 'Text' }
+	],
 	outputs: [],
 	action: () => {},
